+++ src/components/HorizontalBarChart/HorizontalBarChart.tsx
@@ -1,16 +1,18 @@
-import React from 'react';
+import React, {ReactNode} from 'react';
 
+import {TooltipContent} from '../TooltipContent';
 import {ChartContainer} from '../../components/ChartContainer';
 import {usePrefersReducedMotion} from '../../hooks';
 
 import {Chart} from './Chart';
-import type {Series, XAxisOptions} from './types';
+import type {RenderTooltipContentData, Series, XAxisOptions} from './types';
 
 export interface HorizontalBarChartProps {
   series: Series[];
   isAnimated?: boolean;
   isSimple?: boolean;
   isStacked?: boolean;
+  renderTooltipContent?: (data: RenderTooltipContentData) => ReactNode;
   theme?: string;
   xAxisOptions?: XAxisOptions;
 }
@@ -19,6 +21,7 @@ export function HorizontalBarChart({
   isAnimated = true,
   isSimple = false,
   isStacked = false,
+  renderTooltipContent,
   series,
   theme,
   xAxisOptions,
@@ -31,12 +34,29 @@ export function HorizontalBarChart({
 
   const {prefersReducedMotion} = usePrefersReducedMotion();
 
+  function renderTooltip({data}: RenderTooltipContentData) {
+    if (renderTooltipContent != null) {
+      return renderTooltipContent({data});
+    }
+
+    const tooltipData = data.map(({value, label, color}) => {
+      return {
+        label,
+        value: xAxisOptionsForChart.labelFormatter(value),
+        color,
+      };
+    });
+
+    return <TooltipContent data={tooltipData} theme={theme} />;
+  }
+
   return (
     <ChartContainer theme={theme}>
       <Chart
         isAnimated={isAnimated && !prefersReducedMotion}
         isSimple={isSimple}
         isStacked={isStacked}
+        renderTooltipContent={renderTooltip}
         series={series}
         xAxisOptions={xAxisOptionsForChart}
       />