+++ src/components/TooltipContent/index.ts
@@ -1,2 +1,2 @@
 export {TooltipContent} from './TooltipContent';
-export type {TooltipContentProps} from './TooltipContent';
+export type {TooltipContentProps, TooltipData} from './TooltipContent';